Salaries converted to double from int

+++ src/main/java/com/workintech/spring/developer/model/Developer.java
@@ -7,10 +7,10 @@ import org.springframework.stereotype.Component;
 public class Developer {
     private int id;
     private String name;
-    private int salary;
+    private double salary;
     private Experience experience;
 
-    public Developer(int id, String name, int salary, Experience experience) {
+    public Developer(int id, String name, double salary, Experience experience) {
         this.id = id;
         this.name = name;
         this.salary = salary;
@@ -33,7 +33,7 @@ public class Developer {
         this.name = name;
     }
 
-    public int getSalary() {
+    public double getSalary() {
         return salary;
     }
